Consolidate graph extractor
