+++ graphrag/index/operations/extract_entities/typing.py
@@ -5,6 +5,7 @@
 
 from collections.abc import Awaitable, Callable
 from dataclasses import dataclass
+from enum import Enum
 from typing import Any
 
 import networkx as nx
@@ -45,3 +46,14 @@ EntityExtractStrategy = Callable[
     ],
     Awaitable[EntityExtractionResult],
 ]
+
+
+class ExtractEntityStrategyType(str, Enum):
+    """ExtractEntityStrategyType class definition."""
+
+    graph_intelligence = "graph_intelligence"
+    nltk = "nltk"
+
+    def __repr__(self):
+        """Get a string representation."""
+        return f'"{self.value}"'
